Move service down the chain for tests
