+++ sh.exe.stackdump
@@ -1,6 +1,6 @@
 Stack trace:
 Frame     Function  Args
-0081FF94  775A62C4 (00305000, C10F6FE3, 00000000, 00000000)
-0081FFDC  776D0F69 (FFFFFFFF, 776F2EC3, 00000000, 00000000)
-0081FFEC  776D0F34 (00401000, 00305000, 00000000, 57560848)
+0081FF94  775A62C4 (003BF000, 56C3982D, 00000000, 00000000)
+0081FFDC  776D0F69 (FFFFFFFF, 776F2EA7, 00000000, 00000000)
+0081FFEC  776D0F34 (00401000, 003BF000, 00000000, 00000020)
 End of stack trace
