+++ src/Package/Test/Shell/VsAppShellTestSetup.cs
@@ -73,7 +73,6 @@ namespace Microsoft.VisualStudio.R.Package.Test.Shell {
                 .AddService(new REditorSettings(new TestSettingsStorage()))
                 .AddService(new TestImageService())
                 .AddService(new VsEditorSupport(serviceManager))
-                .AddService(new VsEditorViewLocator())
                 .AddWindowsRInterpretersServices()
                 .AddWindowsHostClientServices()
                 .AddWindowsRComponentsServices()
