+++ src/Windows/Editor/Impl/ViewModel/EditorViewLocator.cs
@@ -2,13 +2,15 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Collections.Generic;
+using System.ComponentModel.Composition;
 using System.Linq;
 using Microsoft.Languages.Editor.Controllers.Views;
 using Microsoft.Languages.Editor.Text;
 using Microsoft.VisualStudio.Text;
 
-namespace Microsoft.VisualStudio.R.Package.Shell {
-    internal sealed class VsEditorViewLocator: IEditorViewLocator {
+namespace Microsoft.Languages.Editor.ViewModel {
+    [Export(typeof(IEditorViewLocator))]
+    internal sealed class EditorViewLocator: IEditorViewLocator {
         public IEditorView GetPrimaryView(IEditorBuffer editorBuffer) 
             => GetAllViews(editorBuffer).FirstOrDefault();
